// Raillmg-main/Raillmg-main/client/src/app/shared/model/machineRoll.model.ts
export interface IMachineRoll {
  _id?: string;
  date: string;
  department: string;
  section: string;
  stationTo: string;
  stationFrom: string;
  direction: string;
  mps: number;
  lineNo: string;
  machine: string;
  block: string;
  duration: number;
  section1: string;
  dmd_duration: number;
  avl_duration: number;
  avl_start: string;
  avl_end: string;
  quantum: string;
  deputedSupervisor: string;
  resources: string;
  crew: number;
  loco: number;
  board: string;
  typeOfWork: string;
  ni: string;
  remarks: string;
  approval: string;
  s_tStaff: string;
  tpcStaff: string;
  grant_status: string;
  Avl_status: boolean;
  status: string;
  createdBy: string;
  createdAt: string;
  updatedBy: string;
  updatedAt: string;
  logs: ILog[];
  series: string;
  time_granted: string;
  burst: string;
  caution: [];
  cancelTrain: string;
  integrated: any[];
  km: string;
  Apl_remarks: string;
  Dmd_remarks: string;
  slots: string;
  output: string;
  OPTG_remarks: string;
}

export interface ILog {
  updatedBy: string;
  updatedAt: string;
  field: string;
  oldValue: string;
  newValue: string;
}